Safari-611.1.10/LayoutTests/fast/repaint/outline-auto-with-width-less-than-focus-ring-width-and-offset-repaint.html

<html>
<script>jsTestIsAsync = true;</script>
<script src="../../resources/js-test-pre.js"></script>
<head>
<title>This tests that outline repaint rect for focus ring is correct.</title>
<style>
div {
  position: absolute;
  top: 10px;
  height: 10px;
  width: 10px;
  outline-style: auto;
  outline-width: 1px;
  outline-offset: 2px;
}
</style>
</head>
<body>
<div id=foo></div>
</body>
<script>
setTimeout(
	function() { 
        if (window.internals)
	        internals.startTrackingRepaints();
 		document.getElementById("foo").style.top = "50px";
 		document.body.offsetWidth;

        if (window.internals) {
            shouldNotBe("window.internals.repaintRectsAsText().indexOf('3 5')", "-1");
            shouldNotBe("window.internals.repaintRectsAsText().indexOf('3 45')", "-1");
            internals.stopTrackingRepaints();
        }
        finishJSTest();
	}, 0);
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
