LayoutTests/fast/repaint/outline-auto-with-width-less-than-focus-ring-width-and-offset-repaint-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('3 5') is not -1
PASS window.internals.repaintRectsAsText().indexOf('3 45') is not -1
PASS successfullyParsed is true

TEST COMPLETE
